add ability to favorite a tweet
This was difficult since I wanted to show a favorite/unfavorite on all tweets
for a logged in user. For this to work, I had to do some ugly joins so I knew
which tweet was favorited or not by the logged in user. There is probably a
more clean way. Another way would be to only allow to favorite a tweet when we
only show a single tweet, as in /tweets/:id/show. This pattern will in any
case show up for the retweet feature.